pyproject.toml
[project]
name = "symphony"
version = "0.1.0"
description = "A long-running automation service that polls issue trackers and dispatches coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "httpx>=0.27,<1",
    "pyyaml>=6.0,<7",
    "python-liquid>=1.12,<2",
    "structlog>=24.0,<25",
    "crewai>=0.80,<2",
    "litellm>=1.82.0",
    "crewai-tools>=1.10.1",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "watchdog>=6.0.0",
]

[project.scripts]
symphony = "symphony.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/symphony"]

[dependency-groups]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "respx>=0.21,<1",
    "ruff>=0.8,<1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
